# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ContactCreate": {
        "properties": {
          "category": {
            "enum": [
              "booking_help",
              "service_issue",
              "billing",
              "general",
              "business",
              "privacy_request",
              "provider_question"
            ],
            "title": "Category",
            "type": "string"
          },
          "company": {
            "default": "",
            "maxLength": 0,
            "title": "Company",
            "type": "string"
          },
          "consent_source": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Source"
          },
          "consent_timestamp": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Timestamp"
          },
          "customer_timezone": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Timezone"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "email_consent": {
            "default": false,
            "title": "Email Consent",
            "type": "boolean"
          },
          "language": {
            "anyOf": [
              {
                "maxLength": 16,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "marketing_consent": {
            "default": false,
            "title": "Marketing Consent",
            "type": "boolean"
          },
          "message": {
            "maxLength": 5000,
            "minLength": 10,
            "title": "Message",
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "maxLength": 40,
                "minLength": 7,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "policy_version": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Version"
          },
          "sms_consent": {
            "default": false,
            "title": "Sms Consent",
            "type": "boolean"
          },
          "source_url": {
            "format": "uri",
            "minLength": 1,
            "title": "Source Url",
            "type": "string"
          },
          "subject": {
            "maxLength": 200,
            "minLength": 3,
            "title": "Subject",
            "type": "string"
          },
          "transactional_contact_allowed": {
            "default": true,
            "title": "Transactional Contact Allowed",
            "type": "boolean"
          },
          "utm_campaign": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utm Campaign"
          },
          "utm_content": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utm Content"
          },
          "utm_medium": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utm Medium"
          },
          "utm_source": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utm Source"
          },
          "utm_term": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utm Term"
          }
        },
        "required": [
          "source_url",
          "name",
          "email",
          "category",
          "subject",
          "message"
        ],
        "title": "ContactCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProviderInterestCreate": {
        "properties": {
          "business_name": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Business Name",
            "type": "string"
          },
          "business_website": {
            "anyOf": [
              {
                "format": "uri",
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Website"
          },
          "city": {
            "maxLength": 120,
            "minLength": 2,
            "title": "City",
            "type": "string"
          },
          "company": {
            "default": "",
            "maxLength": 0,
            "title": "Company",
            "type": "string"
          },
          "consent_source": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Source"
          },
          "consent_timestamp": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Timestamp"
          },
          "contact_name": {
            "maxLength": 160,
            "minLength": 2,
            "title": "Contact Name",
            "type": "string"
          },
          "customer_timezone": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Timezone"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "email_consent": {
            "default": false,
            "title": "Email Consent",
            "type": "boolean"
          },
          "language": {
            "anyOf": [
              {
                "maxLength": 16,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "license_details": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Details"
          },
          "marketing_consent": {
            "default": false,
            "title": "Marketing Consent",
            "type": "boolean"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 3000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "phone": {
            "maxLength": 40,
            "minLength": 7,
            "title": "Phone",
            "type": "string"
          },
          "policy_version": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Version"
          },
          "postal_code": {
            "pattern": "^\\d{5}(?:-\\d{4})?$",
            "title": "Postal Code",
            "type": "string"
          },
          "service_categories": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Service Categories",
            "type": "array"
          },
          "sms_consent": {
            "default": false,
            "title": "Sms Consent",
            "type": "boolean"
          },
          "source_url": {
            "format": "uri",
            "minLength": 1,
            "title": "Source Url",
            "type": "string"
          },
          "state": {
            "pattern": "^[A-Z]{2}$",
            "title": "State",
            "type": "string"
          },
          "transactional_contact_allowed": {
            "default": true,
            "title": "Transactional Contact Allowed",
            "type": "boolean"
          },
          "utm_campaign": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utm Campaign"
          },
          "utm_content": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utm Content"
          },
          "utm_medium": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utm Medium"
          },
          "utm_source": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utm Source"
          },
          "utm_term": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utm Term"
          }
        },
        "required": [
          "source_url",
          "business_name",
          "contact_name",
          "email",
          "phone",
          "service_categories",
          "city",
          "state",
          "postal_code"
        ],
        "title": "ProviderInterestCreate",
        "type": "object"
      },
...
      "ServiceRequestCreate": {
        "properties": {
          "address_line1": {
            "maxLength": 240,
            "minLength": 3,
            "title": "Address Line1",
            "type": "string"
          },
          "city": {
            "maxLength": 120,
            "minLength": 2,
            "title": "City",
            "type": "string"
          },
          "company": {
            "default": "",
            "maxLength": 0,
            "title": "Company",
            "type": "string"
          },
          "consent_source": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Source"
          },
          "consent_timestamp": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Timestamp"
          },
          "contact_preference": {
            "enum": [
              "email",
              "phone",
              "text"
            ],
            "title": "Contact Preference",
            "type": "string"
          },
          "customer_timezone": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Timezone"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "email_consent": {
            "default": false,
            "title": "Email Consent",
            "type": "boolean"
          },
          "language": {
            "anyOf": [
              {
                "maxLength": 16,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "marketing_consent": {
            "default": false,
            "title": "Marketing Consent",
            "type": "boolean"
          },
          "name": {
            "maxLength": 160,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "maxLength": 40,
            "minLength": 7,
            "title": "Phone",
            "type": "string"
          },
          "policy_version": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Version"
          },
          "postal_code": {
            "pattern": "^\\d{5}(?:-\\d{4})?$",
            "title": "Postal Code",
            "type": "string"
          },
          "requested_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Date"
          },
          "requested_timing": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Timing"
          },
          "service_description": {
            "maxLength": 4000,
            "minLength": 5,
            "title": "Service Description",
            "type": "string"
          },
          "service_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Id"
          },
          "service_slug": {
            "anyOf": [
              {
                "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Slug"
          },
          "sms_consent": {
            "default": false,
            "title": "Sms Consent",
            "type": "boolean"
          },
          "source_url": {
            "format": "uri",
            "minLength": 1,
            "title": "Source Url",
            "type": "string"
          },
          "state": {
            "pattern": "^[A-Z]{2}$",
            "title": "State",
            "type": "string"
          },
          "transactional_contact_allowed": {
            "default": true,
            "title": "Transactional Contact Allowed",
            "type": "boolean"
          },
          "utm_campaign": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utm Campaign"
          },
          "utm_content": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utm Content"
          },
          "utm_medium": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utm Medium"
          },
          "utm_source": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utm Source"
          },
          "utm_term": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utm Term"
          }
        },
        "required": [
          "source_url",
          "name",
          "email",
          "phone",
          "service_description",
          "address_line1",
          "city",
          "state",
          "postal_code",
          "contact_preference"
        ],
        "title": "ServiceRequestCreate",
        "type": "object"
      },
...
      "SubmissionAccepted": {
        "properties": {
          "downstream_status": {
            "title": "Downstream Status",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "const": "REQUEST_ACCEPTED",
            "default": "REQUEST_ACCEPTED",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "downstream_status"
        ],
        "title": "SubmissionAccepted",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "BREERO API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/contact": {
      "post": {
        "operationId": "contact_api_v1_contact_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Contact",
        "tags": [
          "public-forms"
        ]
      }
    },
...
    "/api/v1/provider-interest": {
      "post": {
        "operationId": "provider_interest_api_v1_provider_interest_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderInterestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Provider Interest",
        "tags": [
          "public-forms"
        ]
      }
    },
...
    "/api/v1/service-requests": {
      "post": {
        "operationId": "service_request_api_v1_service_requests_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Service Request",
        "tags": [
          "public-forms"
        ]
      }
    },
...
  }
}